Fix : 백준 공고별 전체 코드 수정

+++ backend/src/main/java/com/ssafy/backend/domain/analysis/dto/BojLanguagePassCount.java
@@ -7,10 +7,10 @@ import lombok.Getter;
 @Getter
 public class BojLanguagePassCount {
 	private Long languageId;
-	private Integer totalPassCount;
+	private double totalPassCount;
 
 	@QueryProjection
-	public BojLanguagePassCount(Long languageId, Integer totalPassCount) {
+	public BojLanguagePassCount(Long languageId, Double totalPassCount) {
 		this.languageId = languageId;
 		this.totalPassCount = totalPassCount;
 	}
